requirements.txt
SQLAlchemy~=1.3.22


pymediainfo~=5.0.3

scandir~=1.10.0

python-magic~=0.4.18
